+++ client/run_web.py
@@ -1,0 +1,3 @@
+import webapp.web as web
+
+web.run()
