+++ client/webapp/static/style.css
@@ -1,0 +1,45 @@
+body {
+	background: white;
+	color: black;
+	padding: 10px
+	width: 1004px
+}
+
+#header {
+    text-align: center
+}
+
+#stream {
+    width: 512px;
+    height: 288px;
+    align: left;
+    margin: 5px 70px;
+}
+
+#text {
+    width: 480px;
+    height: 288px;
+    align: right;
+    margin: 5px 70px;
+}
+
+form {
+    width: 450px;
+    height: 50px;
+    margin: 5px 90px;
+}
+
+#url_in {
+    width: 350px;
+    height: 30px;
+}
+
+#submit {
+    width: 70px;
+    height: 35px;
+    border: 2px solid;
+    border-radius: 5px;
+    margin: 5px 5px;
+    text-align: center;
+    font-size: 18px;
+}
