+++ client/webapp/templates/index.html
@@ -1,0 +1,20 @@
+<html>
+<head>
+    <title>Video Stream</title>
+    <link rel="stylesheet" href='../static/style.css'/>
+</head>
+<body>
+<div id="header">
+    <h1>Video Streaming Demonstration</h1>
+</div>
+<img src="/home/andrey/PycharmProjects/FaceAnalytics/srv/photo/misha.jpg" id="stream">
+<textarea rows="25" cols="70" id="text">
+            There is nobody there yet :(
+        </textarea>
+<form action="{{ url_for('analyse') }}" method="POST">
+    Web-cam URL:<br>
+    <input type="text" name="camera_url" id="url_in">
+    <input type="submit" value="Go" id="submit">
+</form>
+</body>
+</html>
