--- srv/common/__init__.py
@@ -1,2 +1,0 @@
-from .api import load_image_file, face_locations, batch_face_locations, face_landmarks, face_encodings, compare_faces, \
-    face_distance
