--- srv/common/api.py
@@ -1,120 +1,0 @@
-import dlib
-
-import numpy as np
-import scipy.misc
-
-try:
-    import face_recognition_models
-except:
-    print('I can not find models, ask Andrey about that!')
-    quit()
-
-face_detector = dlib.get_frontal_face_detector()
-
-predictor_68_point_model = face_recognition_models.pose_predictor_model_location()
-pose_predictor_68_point = dlib.shape_predictor(predictor_68_point_model)
-
-predictor_5_point_model = face_recognition_models.pose_predictor_five_point_model_location()
-pose_predictor_5_point = dlib.shape_predictor(predictor_5_point_model)
-
-cnn_face_detection_model = face_recognition_models.cnn_face_detector_model_location()
-cnn_face_detector = dlib.cnn_face_detection_model_v1(cnn_face_detection_model)
-
-face_recognition_model = face_recognition_models.face_recognition_model_location()
-face_encoder = dlib.face_recognition_model_v1(face_recognition_model)
-
-
-def _rect_to_css(rect):
-    return rect.top(), rect.right(), rect.bottom(), rect.left()
-
-
-def _css_to_rect(css):
-    return dlib.rectangle(css[3], css[0], css[1], css[2])
-
-
-def _trim_css_to_bounds(css, image_shape):
-    return max(css[0], 0), min(css[1], image_shape[1]), min(css[2], image_shape[0]), max(css[3], 0)
-
-
-def face_distance(face_encodings, face_to_compare):
-    if len(face_encodings) == 0:
-        return np.empty((0))
-
-    return np.linalg.norm(face_encodings - face_to_compare, axis=1)
-
-
-def load_image_file(file, mode='RGB'):
-    return scipy.misc.imread(file, mode=mode)
-
-
-def _raw_face_locations(img, number_of_times_to_upsample=1, model='hog'):
-    if model == 'cnn':
-        return cnn_face_detector(img, number_of_times_to_upsample)
-    else:
-        return face_detector(img, number_of_times_to_upsample)
-
-
-def face_locations(img, number_of_times_to_upsample=1, model='hog'):
-    if model == 'cnn':
-        return [_trim_css_to_bounds(_rect_to_css(face.rect), img.shape) for face in
-                _raw_face_locations(img, number_of_times_to_upsample, 'cnn')]
-    else:
-        return [_trim_css_to_bounds(_rect_to_css(face), img.shape) for face in
-                _raw_face_locations(img, number_of_times_to_upsample, model)]
-
-
-def batch_face_locations(images, number_of_times_to_upsample=1, batch_size=128):
-    def convert_cnn_detections_to_css(detections):
-        return [_trim_css_to_bounds(_rect_to_css(face.rect), images[0].shape) for face in detections]
-
-    raw_detections_batched = _raw_face_locations_batched(images, number_of_times_to_upsample, batch_size)
-
-    return list(map(convert_cnn_detections_to_css, raw_detections_batched))
-
-
-def face_landmarks(face_image, face_locations=None):
-    landmarks = _raw_face_landmarks(face_image, face_locations)
-    landmarks_as_tuples = [[(p.x, p.y) for p in landmark.parts()] for landmark in landmarks]
-
-    # For a definition of each point index, see https://cdn-images-1.medium.com/max/1600/1*AbEg31EgkbXSQehuNJBlWg.png
-    return [{
-        'chin': points[0:17],
-        'left_eyebrow': points[17:22],
-        'right_eyebrow': points[22:27],
-        'nose_bridge': points[27:31],
-        'nose_tip': points[31:36],
-        'left_eye': points[36:42],
-        'right_eye': points[42:48],
-        'top_lip': points[48:55] + [points[64]] + [points[63]] + [points[62]] + [points[61]] + [points[60]],
-        'bottom_lip': points[54:60] + [points[48]] + [points[60]] + [points[67]] + [points[66]] + [points[65]] + [
-            points[64]]
-    } for points in landmarks_as_tuples]
-
-
-def face_encodings(face_image, known_face_locations=None, num_jitters=1):
-    raw_landmarks = _raw_face_landmarks(face_image, known_face_locations, model='small')
-
-    return [np.array(face_encoder.compute_face_descriptor(face_image, raw_landmark_set, num_jitters)) for
-            raw_landmark_set in raw_landmarks]
-
-
-def compare_faces(known_face_encodings, face_encoding_to_check, tolerance=0.6):
-    return list(face_distance(known_face_encodings, face_encoding_to_check) <= tolerance)
-
-
-def _raw_face_locations_batched(images, number_of_times_to_upsample=1, batch_size=128):
-    return cnn_face_detector(images, number_of_times_to_upsample, batch_size=batch_size)
-
-
-def _raw_face_landmarks(face_image, face_locations=None, model='large'):
-    if face_locations is None:
-        face_locations = _raw_face_locations(face_image)
-    else:
-        face_locations = [_css_to_rect(face_location) for face_location in face_locations]
-
-    pose_predictor = pose_predictor_68_point
-
-    if model == 'small':
-        pose_predictor = pose_predictor_5_point
-
-    return [pose_predictor(face_image, face_location) for face_location in face_locations]
