--- srv/common/config/config.ini
@@ -1,8 +1,0 @@
-[DEFAULT]
-confidence: 0.75
-descriptions_dir: ../video_processing/tmp/description
-detected_faces_dir: ../video_processing/tmp/faces
-known_people_dir : ../photo
-models_dir : ../models
-model : ../models/res10_300x300_ssd_iter_140000.caffemodel
-prototxt : ../models/deploy.prototxt
