--- srv/common/config/config_parser.py
@@ -1,8 +1,0 @@
-import configparser
-
-parser = configparser.ConfigParser()
-
-
-def parse_default(config_path='../common/config/config.ini'):
-    parser.read(config_path)
-    return parser['DEFAULT']
