+++ srv/config/config.ini
@@ -1,0 +1,8 @@
+[DEFAULT]
+confidence : 0.75
+descriptions_dir : ../face_description/tmp/description
+detected_faces_dir : ../face_description/tmp/faces
+known_people_dir : ../face_description/known_faces
+models_dir : ../face_description/models
+model : ../face_description/models/res10_300x300_ssd_iter_140000.caffemodel
+prototxt : ../face_description/models/deploy.prototxt
