+++ srv/config/config_parser.py
@@ -1,0 +1,8 @@
+import configparser
+
+parser = configparser.ConfigParser()
+
+
+def parse_default(config_path='../config/config.ini'):
+    parser.read(config_path)
+    return parser['DEFAULT']
