+++ srv/face_description/run_descriptor.py
@@ -1,0 +1,3 @@
+from face_description import face_descriptor
+
+face_descriptor.start()
