--- srv/flask_api/run.py
@@ -1,3 +1,0 @@
-from flask_api import flask_streaming_api
-
-flask_streaming_api.run()
