+++ srv/flask_api/run_flask.py
@@ -1,0 +1,3 @@
+from flask_api import flask_streaming_api
+
+flask_streaming_api.run()
