--- srv/models/__init__.py
@@ -1,3 +1,0 @@
-
-
-
