--- srv/video_processing/run.py
@@ -1,3 +1,0 @@
-from video_processing import face_descriptor
-
-face_descriptor.start()
